+++ example/lib/customs/pickers/multi_tabs_assets_picker.dart
@@ -129,6 +129,7 @@ class _MultiTabAssetPickerState extends State<MultiTabAssetPicker> {
   Widget selectedAssetsListView(BuildContext context) {
     return Expanded(
       child: ListView.builder(
+        shrinkWrap: true,
         physics: const BouncingScrollPhysics(),
         padding: const EdgeInsets.symmetric(horizontal: 8.0),
         scrollDirection: Axis.horizontal,
@@ -295,7 +296,7 @@ class MultiTabAssetPickerBuilder extends DefaultAssetPickerBuilderDelegate {
           child: Container(
             height: appBarItemHeight,
             constraints: BoxConstraints(
-              maxWidth: MediaQuery.of(context).size.width * 0.5,
+              maxWidth: MediaQuery.sizeOf(context).width * 0.5,
             ),
             padding: const EdgeInsetsDirectional.only(start: 12, end: 6),
             decoration: BoxDecoration(
@@ -425,7 +426,7 @@ class MultiTabAssetPickerBuilder extends DefaultAssetPickerBuilderDelegate {
 
   @override
   AssetPickerAppBar appBar(BuildContext context) {
-    return AssetPickerAppBar(
+    final AssetPickerAppBar appBar = AssetPickerAppBar(
       backgroundColor: theme.appBarTheme.backgroundColor,
       centerTitle: true,
       title: Semantics(
@@ -443,12 +444,12 @@ class MultiTabAssetPickerBuilder extends DefaultAssetPickerBuilderDelegate {
         ],
       ),
     );
+    appBarPreferredSize ??= appBar.preferredSize;
+    return appBar;
   }
 
   @override
   Widget appleOSLayout(BuildContext context) {
-    final AssetPickerAppBar appBarWidget = appBar(context);
-
     Widget layout(BuildContext context) {
       return Stack(
         children: <Widget>[
@@ -469,7 +470,7 @@ class MultiTabAssetPickerBuilder extends DefaultAssetPickerBuilderDelegate {
               ),
             ],
           ),
-          appBarWidget,
+          appBar(context),
         ],
       );
     }